app/requirements.txt
flask
pymongo
cassandra-driver
pandas
matplotlib
seaborn
statsmodels
scikit-learn
joblib